api/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
Faker
requests
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
argon2-cffi
slowapi